apps/file-service/pyproject.toml
[project]
name = "file-service"
version = "0.1.0"
description = "File management service with gRPC interface"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "grpcio-health-checking>=1.76.0",
    "protobuf>=6.31.1,<7.0.0",
    "pymongo>=4.16.0",
    "boto3>=1.35.0",
    "python-dotenv>=1.2.1,<2.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation>=0.60b1",
    "opentelemetry-instrumentation-requests>=0.60b1",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.9",
    "pyjwt>=2.8.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build.targets.wheel]
packages = ["file_service", "user", "auth", "file"]
